pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "EarthquakeMetadata"
version = "0.1.0"
description = "Backend for the earthquake web-app"
readme = "README.md"
authors = [
    {name = "Carrow Morris-Wiltshre", email = "c.morris-wiltshire@newcastle.ac.uk"},
    {name = "Aleksandra Zaforemska", email = "a.zaforemska2@newcastle.ac.uk"},
    {name = "Kristina Wolf", email = "k.wolf2@newcastle.ac.uk"}
]
license = { file = "LICENSE" }
requires-python = ">=3.10,<3.11"
dependencies = [
    "cachetools==5.3.0",
    "certifi==2023.5.7",
    "charset-normalizer==3.1.0",
    "future==0.18.3",
    "google-api-core==2.11.0",
    "google-auth==2.19.0",
    "google-cloud-core==2.3.2",
    "google-cloud-storage==2.9.0",
    "google-crc32c==1.5.0",
    "google-resumable-media==2.5.0",
    "googleapis-common-protos==1.59.0",
    "idna==3.4",
    "numpy==1.24.3",
    "pandas==2.0.1",
    "panoptes-client==1.6.0",
    "Pillow==9.5.0",
    "protobuf==4.23.1",
    "pyasn1==0.5.0",
    "pyasn1-modules==0.3.0",
    "python-dateutil==2.8.2",
    "python-magic==0.4.27",
    "pytz==2023.3",
    "redo==2.0.4",
    "requests>=2.28.0",
    "rsa==4.9",
    "six==1.16.0",
    "tzdata==2023.3",
    "urllib3==1.26.16"
] # these are the dependencies that will be installed in the virtual environment .venv directory

[tool.setuptools]
packages = ["package"] # this line specifies the package location